Updated memory file system to support directories.
Adjusted files that are greater than max cache size to now still at least have a record.

Reorganized the code just a bit to be less all over the place.

environmental variables are no longer evaluated with every file on the file system.

Logic to populate the file system is now in the ... file system.

Added a few test images for the default site.